fixing schema, dags and image transformation

+++ utils/datatypes.py
@@ -8,7 +8,7 @@ Dimensions = Tuple[int, int]
 
 # Polars dataframe schema
 df_schema = {
-    "index": pl.UInt16,
+    "index": pl.UInt32,
     "depth": pl.UInt8,
     "file": pl.String,
     "result": pl.String,
